Restructured the assets project by adding some packages. Added the empty AssetGenerationController class.
