+++ assets/src/main/java/com/adversespaceloneliness/game/assets/Main.java
@@ -1,14 +1,10 @@
 package com.adversespaceloneliness.game.assets;
 
-import org.apache.commons.io.FileUtils;
-
-import java.io.File;
-import java.util.Arrays;
+import com.adversespaceloneliness.game.assets.generation.AssetGenerationController;
 
 public class Main {
 
     public static void main(String[] args) {
-        System.out.println("Assets' main called");
-        System.out.println( Arrays.toString(FileUtils.listFiles(new File("assets/"), null, true).toArray()));
+        AssetGenerationController controller = new AssetGenerationController();
     }
 }
