+++ assets/src/main/java/com/adversespaceloneliness/game/assets/generation/AssetGenerationController.java
@@ -1,0 +1,8 @@
+package com.adversespaceloneliness.game.assets.generation;
+
+public class AssetGenerationController {
+
+    public static final String ASSET_DIRECTORY = "assets/";
+    public static final String RAW_DIRECTORY = ASSET_DIRECTORY + "raw/";
+    public static final String GENERATED_DIRECTORY = ASSET_DIRECTORY + "generated/";
+}
